--- src/main/java/dev/cachaguercu/proyecto4/controllers/GhostController.java
@@ -1,5 +1,0 @@
-package dev.cachaguercu.proyecto4.controllers;
-
-public class GhostController {
-    
-}
