--- src/main/java/dev/cachaguercu/proyecto4/models/GhostBusterModel.java
@@ -1,5 +1,0 @@
-package dev.cachaguercu.proyecto4.models;
-
-public class GhostBusterModel {
-    
-}
