--- src/main/java/dev/cachaguercu/proyecto4/models/GhostModel.java
@@ -1,5 +1,0 @@
-package dev.cachaguercu.proyecto4.models;
-
-public class GhostModel {
-    
-}
