+++ src/main/java/dev/cachaguercus/proyecto4/controllers/GhostBusterController.java
@@ -1,0 +1,5 @@
+package dev.cachaguercus.proyecto4.controllers;
+
+public class GhostBusterController {
+
+}
