+++ src/main/java/dev/cachaguercus/proyecto4/controllers/GhostController.java
@@ -1,0 +1,5 @@
+package dev.cachaguercus.proyecto4.controllers;
+
+public class GhostController {
+
+}
