+++ src/main/java/dev/cachaguercus/proyecto4/models/GhostBusterModel.java
@@ -1,0 +1,5 @@
+package dev.cachaguercus.proyecto4.models;
+
+public class GhostBusterModel {
+
+}
