--- src/test/java/dev/cachaguercu/proyecto4/controllers/GhostBusterControllerTest.java
@@ -1,4 +1,0 @@
-package dev.cachaguercu.proyecto4.controllers;
-
-public class GhostBusterControllerTest {
-}
