--- src/test/java/dev/cachaguercu/proyecto4/controllers/GhostControllerTest.java
@@ -1,4 +1,0 @@
-package dev.cachaguercu.proyecto4.controllers;
-
-public class GhostControllerTest {
-}
