--- src/test/java/dev/cachaguercu/proyecto4/models/GhostModelTest.java
@@ -1,4 +1,0 @@
-package dev.cachaguercu.proyecto4.models;
-
-public class GhostModelTest {
-}
