+++ src/test/java/dev/cachaguercus/proyecto4/controllers/GhostBusterControllerTest.java
@@ -1,0 +1,4 @@
+package dev.cachaguercus.proyecto4.controllers;
+
+public class GhostBusterControllerTest {
+}
