+++ src/test/java/dev/cachaguercus/proyecto4/controllers/GhostControllerTest.java
@@ -1,0 +1,4 @@
+package dev.cachaguercus.proyecto4.controllers;
+
+public class GhostControllerTest {
+}
