+++ src/test/java/dev/cachaguercus/proyecto4/models/GhostBusterModelTest.java
@@ -1,0 +1,4 @@
+package dev.cachaguercus.proyecto4.models;
+
+public class GhostBusterModelTest {
+}
