+++ src/test/java/dev/cachaguercus/proyecto4/models/GhostModelTest.java
@@ -1,0 +1,4 @@
+package dev.cachaguercus.proyecto4.models;
+
+public class GhostModelTest {
+}
