+++ src/test/java/dev/cachaguercus/proyecto4/test_proyecto4/AppTest.java
@@ -1,4 +1,4 @@
-package dev.cachaguercu.test_proyecto4;
+package dev.cachaguercus.proyecto4.test_proyecto4;
 
 import org.junit.jupiter.api.Test;
